fix: Allow peers to reconnect to group chats using a password
This commit deprecates the tox_group_reconnect groupchat API function and
modifies the tox_group_join function to automatically reconnect to groups
when it's called with a chat_id that designates a group that it's already
a member of.

This allows clients to implement group rejoin functionality that lets peers
rejoin/reconnect to groups with a passed password argument, which may be
necessary if the group password changes while a peer is offline. This fixes
the bug described in #2806.