cleanup: Make websockify output qtox-compatible logging.
At the moment this is just because qtox outputs logs in this format and
now I'm used to it, so I want to unify our log outputs to this format.
We don't really parse this output, but after this change, we could
theoretically parse the websockify log output in the qtox/debugtox log
viewer.