chore: Move one of the 3 freebsd builds to post-submit.
This one takes a long time and never fails because we ignore test
failures (of which it has many). The other 2 freebsd builds actually
check for test failures so are more valuable. We keep this one for now
because it's our own freebsd image build, so we at least want to keep it
building in case the other 2 freebsd builds break (e.g. Cirrus shuts
down or the freebsd vm workflow goes away).