chore: Don't upload ios/macos variants in deploy workflows.
We already put them into an xcframework. There's no strong reason to
keep them as separate assets as well.

Also removed the Makefile for single-file uploads, because it
individually appears in the release assets. The .c files are
intentionally separate asset uploads, but the Makefile is always the
same so can be fetched directly from the repo if needed.